schemas for db fields can now be marked required
This is a no-op, but sometimes it is convenient to allow this because a
schema may be used both as a property on an object (where it can be
required) or a top-level schema (where required is not permitted and would
cause the ajv compiler to fail if it wasn't removed).

Require fluent-schema; json schema no longer an option. Consistency.