catch error code 10 regardless of the error message
Firestore updated their error message which caused our code to no longer
catch it. The new error message is "Optimstic transaction was aborted"
(including this typo in "optimistic").

This commit changes to only look at the error code. It also adds logging
of errors so there's more information in case of future issues.